Rename _parse_utils to cvrp
