+++ pycvrplib/read.py
@@ -1,6 +1,6 @@
 from typing import List, Optional, Union
 
-from ._parse_utils import Instance, parse_cvrp
+from .cvrp import Instance, parse_cvrp
 from .solution import parse_solution
 from .utils import find_set, is_vrptw, parse_instance_name
 from .vrptw import VRPTW, parse_vrptw
